Refactor computeDataSourceCredits
